+++ ouzel/gui/FTFont.cpp
@@ -20,7 +20,7 @@
 			{
 			}
 
-			ouzel::FTFont::FTFont(const std::string & filename, int pt, UTFChars flag) {
+			ouzel::FTFont::FTFont(const std::string & filename, int16_t pt, UTFChars flag) {
 				if (!parseFont(filename, pt, flag))
 				{
 					Log(Log::Level::ERR) << "Failed to parse font " << filename;
@@ -30,7 +30,7 @@
 
 			}
 
-			bool FTFont::parseFont(const std::string & filename, int pt, UTFChars flag)
+			bool FTFont::parseFont(const std::string & filename, int16_t pt, UTFChars flag)
 			{
 				FT_Library library;
 				FT_Face	   face;
@@ -71,14 +71,14 @@
 
 				if (flag && UTFChars::ASCII)
 				{
-					for (int i = 32; i < 127; i++)
+					for (uint16_t i = 32; i < 127; i++)
 					{
 						glyphs.push_back(i);
 					}
 				}
 				if (flag && UTFChars::ASCIIPLUS)
 				{
-					for (int i = 128; i < 256; i++)
+					for (uint16_t i = 128; i < 256; i++)
 					{
 						glyphs.push_back(i);
 					}
@@ -99,32 +99,32 @@
 					}
 					FT_Bitmap bitmap = face->glyph->bitmap;
 					CharDescriptor nd;
-					nd.xAdvance = face->glyph->metrics.horiAdvance>>6;
-					nd.height = bitmap.rows;
-					nd.width = bitmap.width;
-					nd.xOffset = face->glyph->metrics.horiBearingX>>6;
-					nd.yOffset = (face->glyph->metrics.height - face->glyph->metrics.horiBearingY)>>6;
+					nd.xAdvance = (int16_t)face->glyph->metrics.horiAdvance>>6;
+					nd.height = (int16_t)bitmap.rows;
+					nd.width = (int16_t)bitmap.width;
+					nd.xOffset = (int16_t)face->glyph->metrics.horiBearingX>>6;
+					nd.yOffset = (int16_t)(face->glyph->metrics.height - face->glyph->metrics.horiBearingY)>>6;
 
 					std::vector<uint8_t> current_buffer((bitmap.rows * bitmap.width));
 					memcpy(current_buffer.data(), bitmap.buffer, (bitmap.rows * bitmap.width));
 
-					glyph_to_bitmap_data.emplace(c, std::make_pair(Size2(bitmap.width, bitmap.rows), current_buffer));
+					glyph_to_bitmap_data.emplace(c, std::make_pair(Size2((float)bitmap.width, (float)bitmap.rows), current_buffer));
 					chars.emplace(c, nd);
-					height = height < bitmap.rows ? bitmap.rows : height;
-					width += bitmap.width;
+					height = height < (int16_t)bitmap.rows ? (int16_t)bitmap.rows : height;
+					width += (int16_t)bitmap.width;
 				}
 				std::vector<std::vector<uint8_t> > scanlines(height, std::vector<uint8_t>());
 				int x = 0;
 				for (const auto &c : glyph_to_bitmap_data)
 				{
-					int char_height = c.second.first.height();
-					int char_width = c.second.first.width();
-					chars.at(c.first).x = x;
+					uint16_t char_height = (uint16_t)c.second.first.height();
+					uint16_t char_width = (uint16_t)c.second.first.width();
+					chars.at(c.first).x = (int16_t)x;
 					chars.at(c.first).y = 0;
 					x += char_width;
 
 
-					int extra_rows = height - char_height;
+					uint16_t extra_rows = height - char_height;
 					chars.at(c.first).yOffset += extra_rows;
 					size_t extra_space_size = extra_rows * char_width;
 					std::vector<uint8_t> new_char_buffer(extra_space_size + char_height * char_width, 0x00);
@@ -156,7 +156,7 @@
 				texture = filename;
 				pages = 1;
 				lineHeight = pt;
-				kernCount = kern.size();
+				kernCount = static_cast<uint16_t>(kern.size());
 				base = 0;
 				return false;
 			}
